Add runtime mechanism in embeddings core, refactor code [WIP]
