+++ lantern_embeddings/src/lib.rs
@@ -1,5 +1,5 @@
 use csv::Writer;
-use lantern_embeddings_core::clip;
+use lantern_embeddings_core::core::get_runtime;
 use lantern_logger::{LogLevel, Logger};
 use lantern_utils::{append_params_to_uri, get_full_table_name, quote_ident};
 use rand::Rng;
@@ -161,8 +161,8 @@ fn embedding_worker(
     let handle = std::thread::spawn(move || {
         let mut count: u64 = 0;
         let model = &args.model;
-        let data_path = &args.data_path;
         let mut start = Instant::now();
+        let runtime = get_runtime(&args.runtime, None, &args.runtime_params)?;
 
         while let Ok(rows) = rx.recv() {
             if is_canceled.is_some() && *is_canceled.as_ref().unwrap().read().unwrap() {
@@ -190,8 +190,7 @@ fn embedding_worker(
                 continue;
             }
 
-            let response_embeddings =
-                clip::process(&model, &input_vectors, None, data_path.as_deref(), false);
+            let response_embeddings = runtime.process(&model, &input_vectors);
 
             if let Err(e) = response_embeddings {
                 anyhow::bail!("{}", e);
@@ -514,6 +513,7 @@ pub fn show_available_models(
 ) -> AnyhowVoidResult {
     let logger = logger.unwrap_or(Logger::new("Lantern Embeddings", LogLevel::Info));
     logger.info("Available Models\n");
-    logger.print_raw(&clip::get_available_models(args.data_path.as_deref()).0);
+    let runtime = get_runtime(&args.runtime, None, &args.runtime_params)?;
+    logger.print_raw(&runtime.get_available_models().0);
     Ok(())
 }
