+++ lantern_embeddings_core/Cargo.toml
@@ -23,3 +23,5 @@ futures = "0.3.28"
 sysinfo = "0.29.11"
 nvml-wrapper = "0.9.0"
 isahc = "1.7.2"
+serde = { version = "1.0", features = ["derive"] }
+serde_json = "1.0.111"
